+++ docs/src/app/layout.tsx
@@ -2,7 +2,7 @@ import { type Metadata } from "next";
 import { Inter } from "next/font/google";
 import localFont from "next/font/local";
 import { Providers } from "@/app/providers";
-import { Layout } from "@/components/Layout";
+import { Layout } from "@/components/layout";
 import { Analytics } from "@vercel/analytics/react";
 import clsx from "clsx";