+++ docs/src/components/layout.tsx
@@ -3,11 +3,11 @@
 import { useEffect, useState } from "react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { Hero } from "@/components/Hero";
-import { Logo, Logomark } from "@/components/Logo";
-import { MobileNavigation } from "@/components/MobileNavigation";
-import { Navigation } from "@/components/Navigation";
-import { Search } from "@/components/Search";
+import { Hero } from "@/components/hero";
+import { Logo, Logomark } from "@/components/logo";
+import { MobileNavigation } from "@/components/mobile-navigation";
+import { Navigation } from "@/components/navigation";
+import { Search } from "@/components/search";
 import clsx from "clsx";
 
 function GitHubIcon(props: React.ComponentPropsWithoutRef<"svg">) {
@@ -35,7 +35,7 @@ function Header() {
   return (
     <header
       className={clsx(
-        "sticky top-0 z-50 flex flex-none flex-wrap items-center justify-between bg-white px-4 py-5 shadow-md shadow-neutral-900/5 transition duration-500 dark:shadow-none sm:px-6 lg:px-8",
+        "sticky top-0 z-50 flex flex-none flex-wrap items-center justify-between bg-white px-4 py-5 shadow-md shadow-neutral-900/5 transition duration-500 sm:px-6 lg:px-8 dark:shadow-none",
         isScrolled
           ? "dark:bg-neutral-900/95 dark:backdrop-blur dark:[@supports(backdrop-filter:blur(0))]:bg-neutral-900/75"
           : "dark:bg-transparent",
@@ -47,7 +47,7 @@ function Header() {
       <div className="relative flex flex-grow basis-0 items-center">
         <Link href="/" aria-label="Home page">
           <Logomark className="h-9 w-9 lg:hidden" />
-          <Logo className="hidden h-9 w-auto fill-neutral-700 dark:fill-neutral-100 lg:block" />
+          <Logo className="hidden h-9 w-auto fill-neutral-700 lg:block dark:fill-neutral-100" />
         </Link>
       </div>
       <div className="-my-5 mr-6 sm:mr-8 md:mr-0">