+++ docs/src/components/navigation.tsx
@@ -22,7 +22,7 @@ export function Navigation({
             </h2>
             <ul
               role="list"
-              className="mt-2 space-y-2 border-l-2 border-neutral-100 dark:border-neutral-800 lg:mt-4 lg:space-y-4 lg:border-neutral-200"
+              className="mt-2 space-y-2 border-l-2 border-neutral-100 lg:mt-4 lg:space-y-4 lg:border-neutral-200 dark:border-neutral-800"
             >
               {section.links.map((link) => (
                 <li key={link.href} className="relative">