+++ packages/builder-react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @coltorapps/builder-react
 
+## 0.2.0
+
+### Minor Changes
+
+- Implemented singular entity rendering in builder and interpreter stores by removing context.
+
+### Patch Changes
+
+- Updated dependencies
+  - @coltorapps/builder@0.2.0
+
 ## 0.1.3
 
 ### Patch Changes